# Prreferred constructor performance test